refactor: transfer tasks endpoints into separate file
